Moved the test to the testing folder,
Also adds the allIsSet function as a utility folder
